HLRest: Move xPackInfo() to xPack().info()
Originally I put the X-Pack info object into the top level rest client
object. I did that because we thought we'd like to squash `xPack` from
the name of the X-Pack APIs now that it is part of the default
distribution. We still kind of want to do that, but at least for now we
feel like it is better to keep the high level rest client aligned with
the other language clients like C# and Python. This shifts the X-Pack
info API to align with its json spec file.
